+++ core/docs/changelog.md
@@ -2,6 +2,9 @@
 ### Fixes
 * Improved error reporting for typos in service ids to pause/unpause service
 
+### Changes
+* Switch the API container to storing its data in the enclave data volume, rather than in the bindmounted enclave data dirpath on the host machine
+
 # 1.45.4
 ### Features
 * Exposed service pause/unpause functionality on the API container